Improve handling of video-player-placeholder element
This commit simplifies how the element named above and its descendants
are managed by the script. It makes it easier to follow on what changes
have been made to the said element.

Also a side effect of this change is that it fixes the issue of the nav
element not being added to its respective position as a
child of
WatchVod.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>