Fix CSS style issues
The team names in the nav element weren't positioned properly due
to a conflicting rule in app.css. This commit fixes that.

The team logos on the side bar didn't match with how they were arranged
on the bar below the video. This commit aims to ensure the arrangement
is similar on both places.

This commit fixes the issue of there being a gap between the side bar
and the main content on the left.

Also it ensures the side bar's background color matches the rest of the
webpage's color.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>