Reduce latency of gathering event details
Due to some changes in the javascript behind lolesports, the previous
method of repeating the XHR call for retrieving event details resulted
in the content script loading before relevant data was ready.

The change in stats.js helps introduce a delay that allows for the data
to be retrieved and stored in the local storage.

The new_background.js will eventually replace the previous background.js

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>