Implement templates for basic player stats
These changes will allow for the display of the players' kill participation
and kda.

Changed the helper in teamStats from unless to if because it makes more
sense to have the loading spinner shown when showSpinner is true.

Added a custom handlebar helper that is an if statment that checks for
equality.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>